TwoCoupledPseudoMotor Controller
First version of the two coupled pseudo motor controller.This
pseudo motor is mapped to two identical motors. Any action applied
to the pseudo motor (move, set, etc...) is applied to both motors
with the same values. The position of both motors must be the same
whithin a certain tolerance. If the tolerance is exceeded, an
exception is raised. Nevertheless, one can unset the validation
of the tolerance by setting the tolerance to -1.