Update BL09 pseudos to deal with new variable offset
After bumping of M1, BL09 pseudomotors and pseudocounters for
Energy have to take into account a variable offset which is calculated
following the equation of a straight line. The offset varies linearly
with the Energy.

Take into account this new offset by using two straight line
coefficients a and b (y = a*x + b)